[frontend] separated tabs (#7279)
