+++ opencti-platform/opencti-front/src/private/Index.tsx
@@ -20,7 +20,7 @@ import useDraftContext from '../utils/hooks/useDraftContext';
 
 const Dashboard = lazy(() => import('./components/Dashboard'));
 const StixObjectOrStixRelationship = lazy(() => import('./components/StixObjectOrStixRelationship'));
-const SearchBulk = lazy(() => import('./components/SearchBulk'));
+const RootSearchBulk = lazy(() => import('@components/SearchBulkContainer'));
 const RootAnalyses = lazy(() => import('./components/analyses/Root'));
 const RootCases = lazy(() => import('./components/cases/Root'));
 const RootEvents = lazy(() => import('./components/events/Root'));
@@ -109,7 +109,7 @@ const Index = ({ settings }: IndexProps) => {
               {/* Search need to be rework */}
               <Route path="/search/*" element={boundaryWrapper(RootSearch)} />
               <Route path="/id/:id" element={boundaryWrapper(StixObjectOrStixRelationship)} />
-              <Route path="/search_bulk" element={boundaryWrapper(SearchBulk)} />
+              <Route path="/search_bulk" element={boundaryWrapper(RootSearchBulk)} />
               <Route path="/analyses/*" element={boundaryWrapper(RootAnalyses)} />
               <Route path="/cases/*" element={boundaryWrapper(RootCases)} />
               <Route path="/events/*" element={boundaryWrapper(RootEvents)} />
